+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ## Breaking Changes
 * MPI is no longer supported as a parallelization mode for NetKet. JAX sharding is now the only supported method for distributed computing.
+* {func}`netket.operator.GraphOperator` is now a function and not a class anymore.
 
 ## New features
 * A new {class}`netket.experimental.driver.VMC_SR` driver, which implements both standard SR and the kernel trick/minSR variant, using an often more efficient implementation, is now available. We advise everyone using SR to switch to this driver [#2007](https://github.com/netket/netket/pull/2007).